requirements.txt
Flask
requests
python-dotenv
markdown
openai
gunicorn